# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "USEDESCK_TEST Api",
        "version": "1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "requestBody": {
                    "request": "auth",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Get a JWT via given credentials.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "description": "Get a JWT via given credentials.",
                                        "type": "string",
                                        "example": "a866b892e50261ac8576"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Get a JWT via given credentials.",
                                            "type": "string",
                                            "example": "long_string"
                                        },
                                        "token_type": {
                                            "description": "Get a JWT via given credentials.",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Get a JWT via given credentials.",
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorisation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Get a JWT via given credentials.",
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Log the user out (Invalidate the token).",
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        },
                                        "data": {
                                            "description": "Log the user out (Invalidate the token).",
                                            "items": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::refresh",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Refresh a token.",
                                            "type": "string",
                                            "example": "long_string"
                                        },
                                        "token_type": {
                                            "description": "Refresh a token.",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Refresh a token.",
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorisation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Refresh a token.",
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Clients list and search",
                "operationId": "getClientList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search_field",
                        "in": "query",
                        "description": "Search field",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "name",
                                "last_name",
                                "phone",
                                "email",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search value",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Clients list",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Meta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Add client",
                "operationId": "App\\Http\\Controllers\\Api\\ClientController::store",
                "requestBody": {
                    "request": "client",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Error in question"
                    },
                    "401": {
                        "description": "Authorisation Error"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{client_id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Show client",
                "operationId": "App\\Http\\Controllers\\Api\\ClientController::show",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Error in question"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update client",
                "operationId": "App\\Http\\Controllers\\Api\\ClientController::update",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "request": "Client",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated"
                    },
                    "400": {
                        "description": "Error in question"
                    },
                    "401": {
                        "description": "Authorisation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete client",
                "operationId": "App\\Http\\Controllers\\Api\\ClientController::destroy",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Authorisation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logs": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Logs opertaion list and search",
                "operationId": "getLogOpertions",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "operation",
                        "in": "query",
                        "description": "operation in list:create, update, delete",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Client model",
                "required": [
                    "name",
                    "last_name"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "description": "Class Client",
                        "example": "Ivan"
                    },
                    "last_name": {
                        "description": "Class Client",
                        "example": "Ivanov"
                    },
                    "created_at": {
                        "description": "Class Client",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Class Client",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "emails": {
                        "title": "Client mailboxes",
                        "description": "Class Client",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "ivanov@usedesc.ru"
                        }
                    },
                    "phones": {
                        "title": "Client phones",
                        "description": "Class Client",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "+7(999)9999999"
                        }
                    }
                },
                "type": "object"
            },
            "ClientEmail": {
                "title": "Email client",
                "description": "model email",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "description": "Class ClientEmail",
                        "example": "ivanov@usedesk.ru"
                    }
                },
                "type": "object"
            },
            "ClientPhone": {
                "title": "Phone client",
                "description": "Phone model",
                "required": [
                    "phone"
                ],
                "properties": {
                    "phone": {
                        "description": "Class ClientPhone",
                        "example": "+7(999)9999999"
                    }
                },
                "type": "object"
            },
            "Links": {
                "title": "Links",
                "properties": {
                    "self": {
                        "type": "string",
                        "example": "link-value"
                    },
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8000/api/clients?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8000/api/clients?page=2"
                    },
                    "prev": {
                        "example": null
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost:8000/api/clients?page=2"
                    }
                },
                "type": "object"
            },
            "Meta": {
                "title": "Meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8000/api/clients"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}